requirements.txt
numpy
pandas
matplotlib
scikit-learn
google-genai
